+++ src/Sentry/ITransactionData.cs
@@ -1,0 +1,9 @@
+﻿namespace Sentry
+{
+    /// <summary>
+    /// Immutable data belonging to a transaction.
+    /// </summary>
+    public interface ITransactionData : ISpanData, ITransactionContext, IEventLike
+    {
+    }
+}