+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changes
 
+- Changed the underlying implementation of `ITransaction` and `ISpan`. `IHub.CaptureTransaction` now takes a `Transaction` instead of `ITransaction`. (#880) @Tyrrrz
 - Add IsParentSampled to TransactionContext (#885) @Tyrrrz
 - Retrieve CurrentVersion for ASP.NET applications (#884) @lucas-zimerman
 
@@ -15,7 +16,7 @@
 - Enrich transactions with more data (#875) @Tyrrrz
 
 ### Fixes
-  - Don't add version prefix in release if it's already set (#877) @Tyrrrz
+- Don't add version prefix in release if it's already set (#877) @Tyrrrz
 
 ## 3.0.8